deploy/installer: Refactor override
